♻️ separate utils from files
